+++ buildSrc/src/main/kotlin/PluginDependencies.kt
@@ -1,0 +1,1 @@
+../../../../plugin/buildSrc/src/main/kotlin/PluginDependencies.kt